pyproject.toml
[project]
name = "adk_agent_template_base"
version = "0.1.0"
description = "A base template for ADK agents with multiple architecture examples."
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "pytest",
    "pytest-asyncio",
    "ruff",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
